+++ CleverTapSDK/CleverTapBuildInfo.h
@@ -1,3 +1,3 @@
 
-#define WR_SDK_REVISION @"40001"
+#define WR_SDK_REVISION @"40100"